Pilot-31 Lots of commit for synchronized learning
Wait(); Notify(); synchroized method;
reproduced deadlock
find a tool to visiualize the JVM thread running status and dump.Cool!

+++ src/main/java/com/kelvin/how2jPractice/APHero.java
@@ -2,7 +2,7 @@ package com.kelvin.how2jPractice;
 
 public class APHero extends Hero implements AP{
 
-	public APHero(String name, float hp, int damage) {
+	public APHero(String name, int hp, int damage) {
 		super(name, hp, damage);
 	}
